Fixed copilot module to work both with rotorcraft and fixedwing firmwares
