+++ conf/firmwares/rotorcraft.makefile
@@ -43,6 +43,7 @@ VPATH += $(PAPARAZZI_HOME)/var/share
 ## COMMON ROTORCRAFT ALL TARGETS (AP + NPS)
 ##
 
+$(TARGET).CFLAGS += -DFIRMWARE=ROTORCRAFT
 $(TARGET).CFLAGS += $(ROTORCRAFT_INC)
 $(TARGET).CFLAGS += -DBOARD_CONFIG=$(BOARD_CFG)
 $(TARGET).CFLAGS += -DPERIPHERALS_AUTO_INIT
